add basic appbar from nextcloud-notes (see README.md)
