+++ app/src/main/AndroidManifest.xml
@@ -45,10 +45,6 @@
             </intent-filter>
         </receiver>
 
-        <activity
-            android:name=".SettingsActivity"
-            android:label="@string/app_name">
-        </activity>
     </application>
 
 </manifest>
