+++ app/src/main/java/com/polar/nextcloudservices/MainActivity.java
@@ -1,12 +1,26 @@
 package com.polar.nextcloudservices;
 
+import static android.view.View.GONE;
+import static android.view.View.VISIBLE;
+import static com.polar.nextcloudservices.settings.SettingsUtils.getBoolPreference;
+
+import android.app.ActivityManager;
+import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
+import android.text.TextUtils;
+import android.util.Log;
 import android.view.View;
 import android.view.Menu;
+import android.view.ViewTreeObserver;
+import android.widget.LinearLayout;
 
 import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.navigation.NavigationView;
 
+import androidx.appcompat.widget.SearchView;
+import androidx.core.view.GravityCompat;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
@@ -15,11 +29,17 @@ import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.polar.nextcloudservices.databinding.ActivityMainBinding;
+import com.polar.nextcloudservices.settings.NotificationServiceConnection;
+import com.polar.nextcloudservices.ui.settings.SettingsFragment;
+
+import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
 
+    private final String TAG = "MainActivity";
     private AppBarConfiguration mAppBarConfiguration;
     private ActivityMainBinding binding;
+    private NotificationServiceConnection mServiceConnection = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -47,6 +67,9 @@ public class MainActivity extends AppCompatActivity {
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_main);
         NavigationUI.setupActionBarWithNavController(this, navController, mAppBarConfiguration);
         NavigationUI.setupWithNavController(navigationView, navController);
+
+        setupToolbars();
+        startNotificationService();
     }
 
     @Override
@@ -61,5 +84,82 @@ public class MainActivity extends AppCompatActivity {
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_main);
         return NavigationUI.navigateUp(navController, mAppBarConfiguration)
                 || super.onSupportNavigateUp();
+    }
+
+
+    public void stopNotificationService() {
+        if(isNotificationServiceRunning()) {
+            Log.i(TAG, "Stopping service");
+            Context context = getApplicationContext();
+            context.stopService(new Intent(context, NotificationService.class));
+        }
+    }
+    public void startNotificationService() {
+        ///--------
+        //Log.d(TAG, "startService: ENTERING");
+        if (!isNotificationServiceRunning() && getBoolPreference(this, "enable_polling",true)) {
+            Log.d(TAG, "Service is not running: creating intent to start it");
+            startService(new Intent(getApplicationContext(), NotificationService.class));
+        }
+    }
+
+    private boolean isNotificationServiceRunning() {
+        ActivityManager manager = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
+        List<?> services = manager.getRunningServices(Integer.MAX_VALUE);
+        return (services.size() > 0);
+    }
+
+    private void updateNotificationServiceStatus(SettingsFragment settings) {
+        ActivityManager manager = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
+        List<ActivityManager.RunningServiceInfo> services = manager.getRunningServices(Integer.MAX_VALUE);
+        if (services.size() == 0) {
+            Log.e(TAG, "Service is not running!");
+            settings.setStatus("Disconnected: service is not running");
+        } else if(mServiceConnection==null){
+            mServiceConnection = new NotificationServiceConnection(settings);
+            bindService(new Intent(getApplicationContext(), NotificationService.class), mServiceConnection, 0);
+        } else {
+            mServiceConnection.updateStatus();
+        }
+    }
+
+    /**
+     * Taken from package it.niedermann.owncloud.notes.main.MainActivity
+     * https://github.com/stefan-niedermann/nextcloud-notes
+     *
+     */
+    private void setupToolbars() {
+        setSupportActionBar(binding.appBarMain.toolbar);
+        //updateCurrentAccountAvatar();
+        binding.appBarMain.homeToolbar.setOnClickListener((v) -> {
+            if (binding.appBarMain.toolbar.getVisibility() == GONE) {
+                //updateToolbars(false);
+            }
+        });
+
+        //binding.appBarMain.launchAccountSwitcher.setOnClickListener((v) -> askForNewAccount(this));
+        binding.appBarMain.menuButton.setOnClickListener((v) -> binding.drawerLayout.openDrawer(GravityCompat.START));
+
+        final LinearLayout searchEditFrame = binding.appBarMain.searchView.findViewById(R.id.search_edit_frame);
+
+        binding.appBarMain.searchView.setOnCloseListener(() -> {
+            if (binding.appBarMain.toolbar.getVisibility() == VISIBLE && TextUtils.isEmpty(binding.appBarMain.searchView.getQuery())) {
+                //updateToolbars(true);
+                return true;
+            }
+            return false;
+        });
+        binding.appBarMain.searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                //refreshLists();
+                return true;
+            }
+        });
     }
 }
