+++ app/src/main/java/com/polar/nextcloudservices/settings/SettingsUtils.java
@@ -1,0 +1,20 @@
+package com.polar.nextcloudservices.settings;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+
+import androidx.preference.PreferenceManager;
+
+public class SettingsUtils {
+
+    public static String getPreference(Context context, String key) {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+        return sharedPreferences.getString(key, "");
+    }
+
+    public static boolean getBoolPreference(Context context, String key, boolean fallback) {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+        return sharedPreferences.getBoolean(key, fallback);
+    }
+
+}
