+++ app/src/main/res/drawable/ic_baseline_menu_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#404040"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M3,18h18v-2L3,16v2zM3,13h18v-2L3,11v2zM3,6v2h18L21,6L3,6z"/>
+</vector>
