+++ app/src/main/res/layout/app_bar_main.xml
@@ -7,17 +7,96 @@
     tools:context=".MainActivity">
 
     <com.google.android.material.appbar.AppBarLayout
+        android:id="@+id/appBar"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="@style/Theme.NextcloudServices.AppBarOverlay">
+        android:layout_height="?attr/actionBarSize"
+        app:elevation="0dp">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
-            app:popupTheme="@style/Theme.NextcloudServices.PopupOverlay" />
+            android:layout_height="match_parent"
+            android:visibility="gone"
+            app:contentInsetStartWithNavigation="0dp"
+            app:navigationIcon="@drawable/ic_arrow_back_grey600_24dp"
+            app:titleMarginStart="0dp"
+            tools:title="@string/app_name">
 
+            <androidx.appcompat.widget.SearchView
+                android:id="@+id/search_view"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content" />
+        </androidx.appcompat.widget.Toolbar>
+
+        <com.google.android.material.card.MaterialCardView
+            android:id="@+id/home_toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/spacer_1x"
+            android:layout_marginTop="@dimen/spacer_1hx"
+            android:layout_marginEnd="@dimen/spacer_1x"
+            android:layout_marginBottom="@dimen/spacer_1hx"
+            app:cardBackgroundColor="@color/appbar"
+            app:cardCornerRadius="@dimen/spacer_1x"
+            app:cardElevation="2dp"
+            app:strokeWidth="0dp">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <androidx.appcompat.widget.AppCompatImageButton
+                    android:id="@+id/menu_button"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:background="?attr/selectableItemBackgroundBorderless"
+                    android:paddingStart="@dimen/spacer_1x"
+                    android:paddingTop="@dimen/spacer_2x"
+                    android:paddingEnd="@dimen/spacer_1x"
+                    android:paddingBottom="@dimen/spacer_2x"
+                    android:tint="?attr/colorAccent"
+                    app:srcCompat="@drawable/ic_baseline_menu_24" />
+
+                <com.google.android.material.textview.MaterialTextView
+                    android:id="@+id/search_text"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="center_vertical"
+                    android:layout_marginStart="@dimen/spacer_1x"
+                    android:layout_marginEnd="@dimen/spacer_1x"
+                    android:layout_weight="1"
+                    android:ellipsize="end"
+                    android:gravity="start"
+                    android:lines="1"
+                    android:textSize="16sp"
+                    tools:text="@string/search_in_all" />
+
+                <androidx.appcompat.widget.AppCompatImageView
+                    android:id="@+id/sorting_method"
+                    android:layout_width="?attr/actionBarSize"
+                    android:layout_height="?attr/actionBarSize"
+                    android:layout_gravity="center_vertical|end"
+                    android:background="?attr/selectableItemBackgroundBorderless"
+                    android:contentDescription="test"
+                    android:padding="@dimen/spacer_2x"
+                    android:tint="?attr/colorAccent"
+                    android:translationX="@dimen/spacer_1x"
+                    app:srcCompat="@drawable/ic_deck" />
+
+                <androidx.appcompat.widget.AppCompatImageView
+                    android:id="@+id/launchAccountSwitcher"
+                    android:layout_width="?attr/actionBarSize"
+                    android:layout_height="?attr/actionBarSize"
+                    android:layout_gravity="center_vertical|end"
+                    android:background="?attr/selectableItemBackgroundBorderless"
+                    android:padding="12dp"
+                    android:tooltipText="@string/manage_accounts"
+                    app:srcCompat="@drawable/ic_account_circle_grey_24dp"
+                    tools:targetApi="o" />
+            </LinearLayout>
+
+        </com.google.android.material.card.MaterialCardView>
     </com.google.android.material.appbar.AppBarLayout>
 
     <include layout="@layout/content_main" />
