+++ app/src/main/res/values/dimens.xml
@@ -7,4 +7,9 @@
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="nav_header_vertical_spacing">8dp</dimen>
     <dimen name="nav_header_height">176dp</dimen>
+
+    <!-- Default screen margins, per the Android Design guidelines. -->
+    <dimen name="spacer_1hx">4dp</dimen>
+    <dimen name="spacer_1x">8dp</dimen>
+    <dimen name="spacer_2x">16dp</dimen>
 </resources>
